Made a new sub component to better handle pokemon types
Moved where the sorting of pokemon types happened
