+++ src/logic/CheckTyping.css
@@ -1,0 +1,6 @@
+.typeRow {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  max-width: 100%;
+}
